as.mcmc.bugs(mcmctools)	R Documentation

Convert a bugs object to an mcmc object

Description

This function converts a bugs object to an mcmc object so that the functions of the coda package may be used to analyze the posterior draws in the bugs object.

Usage

## S3 method for class 'bugs':
as.mcmc(x)

Arguments

x 	 bugs object

Details

None.

Value

If x contains multiple chains, the function returns an mcmc.list object. Otherwise, the function returns an mcmc object.

Note

No notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

bugs in R2WinBUGS

Examples

## Not run: 
## Data object "schools.sim" generated from the examples
## in the bugs function of the R2WinBUGS package.
outmcmc <- as.mcmc(schools.sim)

## Gelman Rubin diagnostics
coda:::gelman.diag(outmcmc)
coda:::mcmc.plot(outmcmc)
## End(Not run)

[Package mcmctools version 1.0 Index]