as.mcmc.rjags(mcmctools)	R Documentation

Convert an rjags object to an mcmc or mcmc.list object.

Description

This is a function to convert and object produced by the jags function in the R2jags package. If the rjags object contains more than one chain, then this function will return an mcmc.list object. Otherwise, the function returns an mcmc object.

Usage

## S3 method for class 'rjags':
as.mcmc(x)

Arguments

x 	 rjags object.

Details

No details.

Value

An mcmc.list or mcmc object.

Note

No further notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

coda package.

Examples

## None yet.

[Package mcmctools version 1.0 Index]