caterpillar(mcmctools)	R Documentation

Side-by-side 'line' plots of the posterior draws of parameters

Description

Plots a point for the mean of posterior simulations along with error bars associated with quantiles fo the posterior distribution.

Usage

caterpillar(mcmcout, params, quantiles = list(), random, collapse=FALSE, plot.names="axis", xlim, col,...)

Arguments

mcmcout 	 a matrix, bugs, mcmc, or mcmc.list object. All objects will be coerced to mcmc.list.

params 	 a character string that identifies which variables in the mcmcout should be plotted. If NULL, all parameter will be plotted.

quantiles 	 list with two elements outer and inner. The outer element of the list should contain the quantiles of the posterior draws that will be plotted as the longer, thinner line. The inner element of the list should contain the quantiles of posterior draws that will be plotted as the shorter, thicker line. If missing, the default is to use list(outer=c(0.025,0.975),inner=c(0.16,0.84)), which corresponds to 95% and 68% credible intervals.

random 	 integer specifying how many parameters from each group will be randomly selected for plotting. This argument is useful when mcmcout has a large number of parameters (e.g., from a hierarchical model). If NULL, all parameters will be plotted.

collapse 	 if TRUE, all parallel chains are collapsed into one chain before plotting. If FALSE, parallel chains are plotted nearly on top of each other with colors as specified in col.

plot.names 	 if 'axis,' then parameter names will be plotted on the axis. If 'above,' then variable names will be plotted above the means for each 'caterpillar.' If any other value, no names will be plotted.

xlim 	 a vector containing the upper and lower limits for the x axis. If missing, the minimum and maximum values found by outer quantiles are used.

col 	 a single value or a vector of values specifying the colors to be used in plotting. If missing and only one parallel chain is contained in mcmcout, then the default is to use col=1. If multiple parallel chains are found in mcmcout, then the default is to use col=seq(nchains) + 1, where nchains is the number of parallel chains.

... 	 further arguments passed to the plotting function.

Details

The function the internal mcmctools function get.parnames.to.plot to match the string params in the names of the variables in mcmcout. Quantiles, as specified in the quantiles argument, are computed for the posterior draws of each variable that matches params. The quantiles are then used to create plots of the posterior intervals of each matched variable. Medians are also plotted as points.

This function produces a plot similar to the plots produced by the coefplot function in the R package arm and the caterpillar plots in the WinBUGS software.

Value

None. Creates a plot.

Note

No notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

mcmcplot

Examples

## Create fake MCMC output
nc <- 10
nr <- 1000
fakemcmc <- lapply(1:3,function(i) matrix(rnorm(nc*nr),nrow=nr))
shift <- matrix(rep(rpois(nc,20),nr),nrow=nr,byrow=TRUE)
fakemcmc <- lapply(fakemcmc,function(x) x + shift)
for (i in seq(along=fakemcmc))
   dimnames(fakemcmc[[i]])[[2]] <- c(paste("alpha[",1:5,"]",sep=""),paste("gamma[",1:5,"]",sep=""))
fakemcmc <- as.mcmc.list(lapply(fakemcmc,mcmc))

## Not run: 
## Caterpillar plots of the fake MCMC output
par(mfrow=c(2,2))
caterpillar(fakemcmc, "alpha")
caterpillar(fakemcmc, "gamma")
caterpillar(fakemcmc, "alpha", collapse=T, plot.names="axis", col="blue")
caterpillar(fakemcmc, "gamma", collapse=T, plot.names="above", col="red")
## End(Not run)

[Package mcmctools version 1.0 Index]