corplot(mcmctools)	R Documentation

Plot of a correlation matrix

Description

Creates an image plot of a correlation matrix where colors of different shades represent differing levels of correlation. One possible use of this function is to plot the correlation between posterior draws of an MCMC run. Patterns in the plot can aid in constructing a more efficient blocking structure for the MCMC algorithm.

Usage

corplot(mat, col = gray(11:0/11), ...)

Arguments

mat 	 correlation matrix.

col 	 colors to be used in the plot.

... 	 currently not used.

Details

No details.

Value

Creates a plot.

Note

No notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

image

Examples

## Not run: 
vals <-c(
 1.00,  0.35, -0.65, -0.66,  0.46,  0.42,
 0.35,  1.00, -0.69, -0.64,  0.40, -0.06,
-0.65, -0.69,  1.00,  0.70, -0.57, -0.11,
-0.66, -0.64,  0.70,  1.00, -0.15, -0.10,
 0.46,  0.40, -0.57, -0.15,  1.00,  0.18,
 0.42, -0.06, -0.11, -0.10,  0.18,  1.00)
Rho <- matrix(vals, 6, 6)
dimnames(Rho) <- list(paste("vec[", 1:6, "]", sep=""), paste("vec[", 1:6, "]", sep=""))
corplot(Rho)
## End(Not run)

[Package mcmctools version 1.0 Index]