denoverplot1(mcmctools)	R Documentation

Plot overlaying densities

Description

Creates a plot containing overlaying kernel density estimates from different MCMC simulations. This function is used in the denoverplot function to produce plots of overlaying densities for parameters in common from two different MCMC simulations.

Usage

denoverplot1(..., main = NULL, sub = NULL, col, lty)

Arguments

... 	 one or more vectors or a list containing one or more vectors to be plotted.

main 	 main title for plot.

sub 	 subtitle for plot.

col 	 one or more colors for the lines. Default is to choose seq(n)+1, where n is the number of arguments in ....

lty 	 types of lines to plot. Defaults to 1.

Details

No details.

Value

Creates a plot.

Note

No further notes.

Author(s)

S. McKay Curtis

References

No References.

See Also

denoverplot

Examples

denoverplot1(rnorm(1000), rnorm(1000))
denoverplot1(list(rgamma(1000,1,1), rgamma(1000,1,1)))

[Package mcmctools version 1.0 Index]