get.parnames.to.plot(mcmctools)	R Documentation

Matches groups of parameters to plot in MCMC diagnostics plots.

Description

Utility function that finds the parameter names to plot in the mcmcplot function. For internal use.

Usage

get.parnames.to.plot(parnames, params, random)

Arguments

parnames 	 parameter names from an MCMC run

params 	 partial parameter names that will be used to determine which subset of parnames will be plotted.

random 	 an integer or NULL. If an integer is specified, the function will select only random number of plots from each parameter group for plotting. If NULL, all parameter names from groups specified in params will be plotted.

Details

No details.

Value

A character vector with parameter names from parnames that match the partial parameter names in params.

Note

No further notes.

Author(s)

S. McKay Curtis

References

~put references to the literature/web site here ~

See Also

mcmcplot

Examples

fakenames <- c(paste("gamma[", 1:5, "]", sep=""), paste("alpha[", 1:5,
"]", sep=""))
get.parnames.to.plot(fakenames, "alpha", random=NULL)
get.parnames.to.plot(fakenames, "alpha", random=3)

[Package mcmctools version 1.0 Index]