mcmcplot1(mcmctools)	R Documentation

Graphs for one parameter from an MCMC run

Description

Creates a graph window containing three different plots—a trace plot, a density plot, and an autocorrelation plot—for one parameter in an MCMC run. This function is used by mcmcplot to construct an html file of MCMC diagnostics.

Usage

mcmcplot1(mcmcout, pmtr, col, lty)

Arguments

mcmcout 	 an object that can be coerced to an mcmc object.

pmtr 	 a string matching one parameter name from mcmcout

col 	 colors for plotting each parallel chain. The default is seq(nchains)+1 where nchains is the number of parallel chains in mcmcout. If there is only one parallel chain, then the default is 1.

lty 	 line types for plotting each parallel chain. The default is 1 for all parallel chains.

Details

No details.

Value

Creates a plot.

Note

Only the first parallel chain is used to create the autocorrelation plot. This function is used by mcmcplot to create html output for all the parameters of an MCMC simulation.

Author(s)

S. McKay Curtis

References

No references.

See Also

mcmcplot

Examples

## Create fake MCMC output
fakemcmc <- sapply(1:5, function(dum) rnorm(1000))
colnames(fakemcmc) <- c("gamma[1,1]", "gamma[1,2]", "gamma[1,3]", "sigma[1]", "sigma[2]")
par(mfrow=c(1,2))
mcmcplot1(fakemcmc, "sigma[1]")
mcmcplot1(fakemcmc, "gamma[1,3]")

[Package mcmctools version 1.0 Index]