denoverplot1(mcmctools)	R Documentation

Overlaying densities for parameters from two MCMC simulations.

Description

This function determines which parameters are in common from two different MCMC simulations and plots overlaying density estimates of the parameters in common. This function can be used in debugging MCMC code by comparing distributions of parameters from the development MCMC code and a reference MCMC simulation (e.g., from WinBUGS).

Usage

denoverplot1(mcmc1, mcmc2, nr.plots, mfrow, mfcol, legend=TRUE, col, lty)

Arguments

mcmc1 	 object that can be coerced to an mcmc object.

mcmc2 	 object that can be coerced to an mcmc object.

nr.plots 	 optional argument to be passed to the mult.fig function for setting up the plotting region. The default value is the number of common parameters in mcmc1 and mcmc2 if legend=FALSE or the number of common parameters in mcmc1 and mcmc2 plus one if legend=TRUE.

mfrow 	 optional argument to be passed to mult.fig for setting up the plotting region.

mfcol 	 optional argument to be passed to mult.fig for setting up the plotting region.

legend 	 if TRUE an extra plot in the plotting region is used as a legend.

col 	 colors for plotting the densites. Default is c(2,3)

lty 	 line types for plotting densities. Default is 1

Details

No details.

Value

Creates a plot.

Note

No further notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

denoverplot

Examples

## Not run: 
## Create fake MCMC output
nc <- 10
nr <- 1000
fakemcmc <- lapply(1:3, function(i) matrix(rnorm(nc*nr),nrow=nr))
shift <- matrix(rep(rpois(nc, 20), nr), nrow=nr, byrow=TRUE)
fakemcmc <- lapply(fakemcmc, function(x) x + shift)
for (i in seq(along=fakemcmc))
   dimnames(fakemcmc[[i]])[[2]] <- c(paste("alpha[", 1:5, "]", sep=""),
paste("gamma[", 1:5, "]", sep=""))
fakemcmc <- as.mcmc.list(lapply(fakemcmc,mcmc))

## Create second fake MCMC output
fakemcmc2 <- lapply(1:3, function(i) matrix(rnorm(nc*nr),nrow=nr))
fakemcmc2 <- lapply(fakemcmc2, function(x) x + shift)
for (i in seq(along=fakemcmc2))
   dimnames(fakemcmc2[[i]])[[2]] <- c(paste("alpha[", 1:5, "]", sep=""),
paste("gamma[", 1:5, "]", sep=""))
fakemcmc2 <- as.mcmc.list(lapply(fakemcmc2, mcmc))

## Plot the fake MCMC output
denoverplot1(fakemcmc, fakemcmc2)
## End(Not run)

[Package mcmctools version 1.0 Index]