denplot(mcmctools)	R Documentation

Produce density plots for parameters in an MCMC simulation on a single plot

Description

This function plots densities from an MCMC simulation on a single plot for all parameters (by default) or those parameters indicated by the params argument.

Usage

denplot(mcmcout, params = NULL, random = NULL, auto.layout = TRUE, col, lty, ...)

Arguments

mcmcout 	 an object that can be coerced to an mcmc or mcmc.list object

params 	 character vector indicating which subset of the parameters to plot. The default is to plot all parameters in mcmcout.

random 	 an integer indicating the maximum number of parameters to randomly select for plotting from each group of parameters as specified by the params argument.

auto.layout 	 automatically creates a plot layout using mult.fig if TRUE.

col 	 colors to be used in plotting the densities. Default is 1:nchains + 1, where nchains is the number of parallel chains in mcmcout.

lty 	 line types to be used in plotting. Defaults to 1.

... 	 further arguments to be passed to the plotting function.

Details

No further details.

Value

Creates a plot.

Note

No further notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

mcmcplot

Examples

## Not run: 
## Create fake MCMC output
nc <- 10
nr <- 1000
fakemcmc <- lapply(1:3,function(i) matrix(rnorm(nc*nr),nrow=nr))
shift <- matrix(rep(rpois(nc,20),nr),nrow=nr,byrow=TRUE)
fakemcmc <- lapply(fakemcmc,function(x) x + shift)
for (i in seq(along=fakemcmc))
   dimnames(fakemcmc[[i]])[[2]] <- c(paste("alpha[", 1:5, "]", sep=""),
paste("gamma[", 1:5, "]", sep=""))
fakemcmc <- as.mcmc.list(lapply(fakemcmc, mcmc))

## Plot densities of the fake MCMC output
denplot(fakemcmc)
denplot(fakemcmc, "gamma")
## End(Not run)

[Package mcmctools version 1.0 Index]