mcmcplot(mcmctools)	R Documentation

Plots for MCMC output in HTML format

Description

Creates an HTML file which will display diagnostic plots (trace, density, autocorrelation) from an MCMC simulation. When the number of parameters in an MCMC simulation is large, viewing all plots in a web browser is much easier than clicking through R graph windows.

Usage

mcmcplot(mcmcout, params=NULL, random=NULL, dir=tempdir(), filename="MCMCoutput", extension="html", title="MCMC Plots", heading, col, lty)

Arguments

mcmcout 	 posterior draws. This argument will be coerced to an mcmc object.

params 	 character vector specifying which subsets of parameters to plot. If NULL, mcmcplot will plot all parameters. Regular expressions are used to strip all numbers and punctuation out of the parameter names to find the parameters that match the character strings in params.

random 	 integer specifying how many parameters from each group will be randomly selected for plotting. This argument is useful when mcmcout has a large number of parameters (e.g., from a hierarchical model). If NULL, mcmcplot will plot all parameters.

dir 	 string containing the directory where the plots and the main html file will be stored.

filename 	 string containing the name of the main html file which will contain code to display each plot produced by mcmcplot.

extension 	 string containing the extension to be used for the html file.

title 	 string containing the title to be included in the html file.

heading 	 string containing the heading to be used for the html file.

col 	 vector of colors. This will determine the colors that will be used to plot each chain in the traceplots and density plots. If missing, mcmcplot will use 1:nchain + 1, where nchain is the number of chains per parameter in mcmcout.

lty 	 vector of line types. This will determine the line types that will be used to plot each chain in the traceplots and density plots. If missing, mcmcplot will use 1 for all line types.

Details

No more details at this time.

Value

Invisibly returns a string containing the path to filename.

Note

No further notes.

Author(s)

S. McKay Curtis

References

No references.

See Also

Plotting functions in the coda package

Examples

## Not run: 
## Create fake MCMC output
nc <- 10
nr <- 1000
fakemcmc <- lapply(1:3, function(i) matrix(rnorm(nc*nr),nrow=nr))
shift <- matrix(rep(rpois(nc, 20), nr), nrow=nr, byrow=TRUE)
fakemcmc <- lapply(fakemcmc, function(x) x + shift)
for (i in seq(along=fakemcmc))
   dimnames(fakemcmc[[i]])[[2]] <- c(paste("alpha[", 1:5, "]", sep=""),
paste("gamma[", 1:5, "]", sep=""))
fakemcmc <- as.mcmc.list(lapply(fakemcmc,mcmc))

## Plot the fake MCMC output
mcmcplot(fakemcmc)
mcmcplot(fakemcmc, "gamma")
## End(Not run)

[Package mcmctools version 1.0 Index]