mcmctools-package(mcmctools)	R Documentation

Sundry tools for evaluating posterior draws from an MCMC simulation

Description

Plotting functions for MCMC output meant to complement the functions in R2WinBUGS and coda.

Details

Package: 	 mcmctools

Type: 	 Package

Version: 	 1.0

Date: 	 2008-11-25

License: 	 GPL (>= 2)

LazyLoad: 	 yes

Author(s)

S. McKay Curtis

Maintainer: S. McKay Curtis <s.mckay.curtis@gmail.com>

References

No references.

See Also

<pkg> ~~

Examples

## None.

[Package mcmctools version 1.0 Index]